Move FAM related code into fam.rs
Move Flexible Array Member related code into dedicated fam.rs.
